requirements.txt
numpy=1.17
snakemake=5.10
pandas=0.25
datashader=0.10
colorcet=2.0
requests=2.22
statsmodels=0.10
pyyaml=5.1
